+++ NNUE2.h
@@ -9,7 +9,9 @@
 #include "Def.h"
 
 #ifdef NNUE_EVALUATION_FUNCTION_2
-void ReadNetwork(void);
+extern BOOL NnueFileLoaded;
+
+BOOL LoadNetwork(const char* NnueFileName);
 
 int NetworkEvaluate(BoardItem* Board);
 #endif // NNUE_EVALUATION_FUNCTION_2